fix: GCC -Werror clean build for Linux debug-werror CI
- Lexer: explicit int64_t for INTEGER token from stoll
- Parser: return params without pessimizing move; nameless dynamic_cast in assignment
- builtins: brace/line breaks for misleading-indentation; ffi sigVals under _WIN32 only;
  time_format tm only on Windows; non-Win wait/kill (void) captures
- REPL clear: cast system() return (warn_unused_result on glibc)

Made-with: Cursor

+++ src/compiler/lexer.cpp
@@ -265,7 +265,7 @@ void Lexer::number() {
             std::string numStr = source_.substr(start_, current_ - start_);
             for (auto it = numStr.begin(); it != numStr.end(); ) { if (*it == '_') it = numStr.erase(it); else ++it; }
             if (numStr.empty()) numStr = "0";
-            addToken(TokenType::INTEGER, std::stoll(numStr));
+            addToken(TokenType::INTEGER, static_cast<int64_t>(std::stoll(numStr)));
         }
     } catch (const std::out_of_range&) {
         throw LexerError("Number literal out of range", line_, column_);
